don't unlink non-existing video.ogv

+++ bin/mplayer.pl
@@ -35,7 +35,7 @@ if ( ! $movie && -e 'media/_editing' ) {
 } elsif ( -f $movie && $movie =~ m{\.og[vg]\b}i ) {
 	my $movie_master = $movie;
 	$movie = base_dir($movie) . '/video.ogv';
-	unlink $movie;
+	unlink $movie if -e $movie;
 	symlink base_name($movie_master), $movie;
 	warn "symlink video.ogv -> $movie_master";
 } else {
